Rename config to constants
